Ajustes nas configurações do projeto provider REST

+++ WebserviceREST-req/src/br/com/fiap/dao/GenericDAO.java
@@ -7,8 +7,8 @@ public interface GenericDAO<T, K> {
 	
 	void cadastrar(T entity);
 	void atualizar(T entity);
-	void excluir(int codigo) throws CodigoInvalidoException;
-	T buscar(int id);
+	void excluir(long codigo) throws CodigoInvalidoException;
+	T buscar(long id);
 	void salvar() throws CommitErrorException;
 	
 }
